moved tests from OrderProductTest to OrderItemTest because we refactor OrderProduct to OrderItem and OrderProduct will disapear soon
